feat(bin): improve "collect"

+++ bin/collect.lua
@@ -15,20 +15,13 @@ local right = turtle.turnRight;
 local forward = turtle.forward;
 local back = turtle.back;
 
--- when front of energy cube
-local function goBackToHomePoint()
-  left();
-  back();
-  left();
-  down();
-  down();
-  down();
+local function printFuel()
+  print('=> Initial Fuel: ', turtle.getFuelLevel(), '/', turtle.getFuelLimit())
 end
 
 local function findEnergyCube()
   for i = 1, 16, 1 do
-    turtle.select(i);
-    local item = turtle.getItemDetail()
+    local item = turtle.getItemDetail(i);
     if item and item.name == ENERGY_CUBE_NAME then
       return i;
     end
@@ -37,41 +30,54 @@ local function findEnergyCube()
   return nil;
 end
 
-local function cleanup()
-  up();
-  up();
-  up();
-  left();
-  left();
-
-  for i = 1, 16, 1 do
-    turtle.select(i);
-    turtle.drop();
-  end
-
-  turtle.select(1)
-
-  left();
-  left();
-  down();
-  down();
-  down();
-end
-
-local function reload()
-  up();
-  up();
+-- when on home
+local function goCleanup()
   up();
   right();
   forward();
   right();
+  up();
+  up();
+end
+
+local function cleanup()
+  for i = 1, 16, 1 do
+    if turtle.getItemCount(i) > 0 then
+      turtle.select(i);
+      turtle.drop();
+    end
+  end
+end
+
+-- when front of ender chest
+local function goBackCleanup()
+  right();
+  down();
+  down();
+  forward();
+  right();
+  down();
+end
+
+-- when on home point
+local function goReload()
+  up();
+  right();
+  forward();
+  forward();
+  right();
+  up();
+  up();
+end
+
+local function reload()
   turtle.dig();
   t.collectItems();
 
   local cubeSlot = findEnergyCube()
   if not cubeSlot then
-    goBackToHomePoint()
-    error('Fatal: cube not found')
+    goBackReload()
+    error('Fatal: item "' .. ENERGY_CUBE_NAME .. '" not found')
   end
 
   turtle.select(cubeSlot);
@@ -87,13 +93,23 @@ local function reload()
   end
 
   turtle.place()
+end
 
+-- when front of energy cube
+local function goBackReload()
+  right();
+  down();
+  down();
+  forward();
+  forward();
+  right();
+  down();
+end
+
+local function goCleanupFromReload()
+  right();
+  forward();
   left();
-  back();
-  left();
-  down();
-  down();
-  down();
 end
 
 local function gatherLeft()
@@ -104,7 +120,7 @@ local function gatherLeft()
 end
 
 local function gatherLine()
-  for i = 1, GREENHOUSE_LENGTH - 1, 1 do
+  for _i = 1, GREENHOUSE_LENGTH - 1, 1 do
     gatherLeft();
     forward();
   end
@@ -112,32 +128,96 @@ local function gatherLine()
   gatherLeft();
 end
 
+local function shouldCleanup()
+  for i = 1, 16, 1 do
+    if turtle.getItemCount(i) > 0 then
+      return true;
+    end
+  end
+
+  return false;
+end
+
+local function waitUnblock()
+  print('=> Turtle collect paused');
+
+  while turtle.detectUp() do
+    os.sleep(2)
+  end
+
+  print('=> Turtle collect resumed');
+end
+
 local function gather()
   gatherLine();
-  left();
-  left();
+  right();
+  right();
+
+  -- for _i = 1, GREENHOUSE_LENGTH - 1, 1 do
+  --   forward();
+  -- end
+
   gatherLine();
-  left();
-  left();
+  right();
+  right();
 end
 
-if turtle.detectUp() then
-  error('turtle blocked')
+local function getNbFreeStack()
+  local counter = 0;
+
+  for i = 1, 16, 1 do
+    if turtle.getItemCount(i) == 0 then
+      counter = counter + 1
+    end
+  end
+
+  return counter;
 end
 
-reload();
-cleanup();
-
-while true do
-  gather();
+local function main()
+  turtle.select(1);
 
   if turtle.detectUp() then
-    error('turtle blocked')
+    waitUnblock();
   end
 
-  cleanup();
+  printFuel();
 
-  if turtle.getFuelLevel() <= ENERGY_THRESHOLD then
-    reload()
+  goReload()
+  reload();
+
+  if shouldCleanup() then
+    goCleanupFromReload();
+    cleanup();
+    goBackCleanup();
+  else
+    goBackReload();
   end
+
+  print('=> Turtle collect started')
+
+  while true do
+    gather();
+
+    if turtle.detectUp() then
+      waitUnblock();
+    end
+
+    if getNbFreeStack() <= 2 then
+      goCleanup();
+      cleanup();
+      goBackCleanup();
+    end
+
+    printFuel();
+    if turtle.getFuelLevel() <= ENERGY_THRESHOLD then
+      goReload()
+      reload();
+      goBackReload();
+    end
+  end
+
+  print('=> Turtle collect stopped');
 end
+
+main();
